mute system move and tweaks, lightweight fix
