+++ Content.Server/Speech/Muting/MutingSystem.cs
@@ -1,0 +1,47 @@
+using Content.Server.Chat.Systems;
+using Content.Server.Popups;
+using Content.Server.Speech.Components;
+using Content.Server.Speech.EntitySystems;
+using Content.Shared.Chat.Prototypes;
+using Content.Shared.Speech;
+
+namespace Content.Server.Speech.Muting
+{
+    public sealed class MutingSystem : EntitySystem
+    {
+        [Dependency] private readonly PopupSystem _popupSystem = default!;
+        public override void Initialize()
+        {
+            base.Initialize();
+            SubscribeLocalEvent<MutedComponent, SpeakAttemptEvent>(OnSpeakAttempt);
+            SubscribeLocalEvent<MutedComponent, EmoteEvent>(OnEmote, before: new[] { typeof(VocalSystem) });
+            SubscribeLocalEvent<MutedComponent, ScreamActionEvent>(OnScreamAction, before: new[] { typeof(VocalSystem) });
+        }
+
+        private void OnEmote(EntityUid uid, MutedComponent component, ref EmoteEvent args)
+        {
+            if (args.Handled)
+                return;
+
+            //still leaves the text so it looks like they are pantomiming a laugh
+            if (args.Emote.Category.HasFlag(EmoteCategory.Vocal))
+                args.Handled = true;
+        }
+
+        private void OnScreamAction(EntityUid uid, MutedComponent component, ScreamActionEvent args)
+        {
+            if (args.Handled)
+                return;
+
+            _popupSystem.PopupEntity(Loc.GetString("speech-muted"), uid, uid);
+            args.Handled = true;
+        }
+
+
+        private void OnSpeakAttempt(EntityUid uid, MutedComponent component, SpeakAttemptEvent args)
+        {
+            _popupSystem.PopupEntity(Loc.GetString("speech-muted"), uid, uid);
+            args.Cancel();
+        }
+    }
+}
