+++ Content.Shared/Drunk/DrunkSystem.cs
@@ -17,13 +17,13 @@ public abstract class SharedDrunkSystem : EntitySystem
         if (!Resolve(uid, ref status, false))
             return;
 
+        if (TryComp<LightweightDrunkComponent>(uid, out var trait))
+            boozePower *= trait.BoozeStrengthMultiplier;
+
         if (applySlur)
         {
             _slurredSystem.DoSlur(uid, TimeSpan.FromSeconds(boozePower), status);
         }
-
-        if (TryComp<LightweightDrunkComponent>(uid, out var trait))
-            boozePower *= trait.BoozeStrengthMultiplier;
 
         if (!_statusEffectsSystem.HasStatusEffect(uid, DrunkKey, status))
         {
